gateway/requirements.txt
openai>=1.14
fastapi
uvicorn[standard]
asyncpg
python-dotenv
